Extract trpc & storage stuff to their respective files
